Created StudentHistory & Timeline component, added props to make it reusable (#13)
Timeline takes an array - "items" — as a prop. the array should contain objects that have a date, description and status number (will update this to be in line with the Student model) when integrating front and backend. Status number is needed as there is different styling for updates regarding students dropping out. 

e.g. items: [
{
  date: 20 Feb 2020,
  description: Joined Madwish
  status: 3 
}
]